requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile '.\requirements.in'
#
cachetools==5.3.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   flask
    #   functions-framework
cloudevents==1.9.0
    # via functions-framework
colorama==0.4.6
    # via click
deprecation==2.1.0
    # via cloudevents
flask==2.2.3
    # via functions-framework
functions-framework==3.3.0
    # via -r .\requirements.in
google-api-core[grpc]==2.11.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.16.1
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-bigquery==3.5.0
    # via -r .\requirements.in
google-cloud-core==2.3.2
    # via google-cloud-bigquery
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.1
    # via google-cloud-bigquery
googleapis-common-protos==1.58.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.51.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.51.1
    # via google-api-core
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
packaging==23.0
    # via
    #   deprecation
    #   google-cloud-bigquery
proto-plus==1.22.2
    # via google-cloud-bigquery
protobuf==4.22.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
python-dateutil==2.8.2
    # via google-cloud-bigquery
requests==2.28.2
    # via
    #   -r .\requirements.in
    #   google-api-core
    #   google-cloud-bigquery
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
urllib3==1.26.14
    # via requests
watchdog==2.2.1
    # via functions-framework
werkzeug==2.2.3
    # via flask
